firmware: move power logic to power.h|c
