+++ firmware/common/hackrf_core.c
@@ -36,6 +36,7 @@
 #include "max5864_target.h"
 #include "platform_gpio.h"
 #include "platform_scu.h"
+#include "power.h"
 #include "spi_bus.h"
 #include "w25q80bv_target.h"
 #ifdef IS_PRALINE
@@ -509,168 +510,6 @@ void pin_setup(void)
 	scu_pinmux(scu->PINMUX_GP_CLKIN, SCU_CLK_IN | SCU_CONF_FUNCTION1);
 
 	sgpio_configure_pin_functions(&sgpio_config);
-}
-
-#ifdef IS_PRALINE
-void enable_1v2_power(void)
-{
-	if (IS_PRALINE) {
-		gpio_set(platform_gpio()->gpio_1v2_enable);
-	}
-}
-
-void disable_1v2_power(void)
-{
-	if (IS_PRALINE) {
-		gpio_clear(platform_gpio()->gpio_1v2_enable);
-	}
-}
-
-void enable_3v3aux_power(void)
-{
-	if (IS_PRALINE) {
-		gpio_clear(platform_gpio()->gpio_3v3aux_enable_n);
-	}
-}
-
-void disable_3v3aux_power(void)
-{
-	if (IS_PRALINE) {
-		gpio_set(platform_gpio()->gpio_3v3aux_enable_n);
-	}
-}
-#endif
-
-void enable_1v8_power(void)
-{
-#ifdef IS_NOT_PRALINE
-	if (IS_NOT_PRALINE) {
-	#ifdef IS_H1_R9
-		if (IS_H1_R9) {
-			gpio_set(platform_gpio()->h1r9_1v8_enable);
-		}
-	#endif
-	#ifdef IS_NOT_H1_R9
-		if (IS_NOT_H1_R9) {
-			gpio_set(platform_gpio()->gpio_1v8_enable);
-		}
-	#endif
-	}
-#endif
-}
-
-void disable_1v8_power(void)
-{
-#ifdef IS_NOT_PRALINE
-	if (IS_NOT_PRALINE) {
-	#ifdef IS_H1_R9
-		if (IS_H1_R9) {
-			gpio_clear(platform_gpio()->h1r9_1v8_enable);
-		}
-	#endif
-	#ifdef IS_NOT_H1_R9
-		if (IS_NOT_H1_R9) {
-			gpio_clear(platform_gpio()->gpio_1v8_enable);
-		}
-	#endif
-	}
-#endif
-}
-
-#ifdef IS_HACKRF_ONE
-static inline void enable_rf_power_hackrf_one(void)
-{
-	const platform_gpio_t* gpio = platform_gpio();
-	uint32_t i;
-
-	/* many short pulses to avoid one big voltage glitch */
-	for (i = 0; i < 1000; i++) {
-		if (detected_platform() == BOARD_ID_HACKRF1_R9) {
-			gpio_set(gpio->h1r9_vaa_disable);
-			gpio_clear(gpio->h1r9_vaa_disable);
-		} else {
-			gpio_set(gpio->vaa_disable);
-			gpio_clear(gpio->vaa_disable);
-		}
-	}
-}
-
-static inline void disable_rf_power_hackrf_one(void)
-{
-	if (detected_platform() == BOARD_ID_HACKRF1_R9) {
-		gpio_set(platform_gpio()->h1r9_vaa_disable);
-	} else {
-		gpio_set(platform_gpio()->vaa_disable);
-	}
-}
-#endif
-
-#ifdef IS_PRALINE
-static inline void enable_rf_power_praline(void)
-{
-	gpio_clear(platform_gpio()->vaa_disable);
-
-	/* Let the voltage stabilize */
-	delay(1000000);
-}
-
-static inline void disable_rf_power_praline(void)
-{
-	gpio_set(platform_gpio()->vaa_disable);
-}
-#endif
-
-#ifdef IS_RAD1O
-static inline void enable_rf_power_rad1o(void)
-{
-	gpio_set(platform_gpio()->vaa_enable);
-
-	/* Let the voltage stabilize */
-	delay(1000000);
-}
-
-static inline void disable_rf_power_rad1o(void)
-{
-	gpio_clear(platform_gpio()->vaa_enable);
-}
-#endif
-
-void enable_rf_power(void)
-{
-#ifdef IS_HACKRF_ONE
-	if (IS_HACKRF_ONE) {
-		enable_rf_power_hackrf_one();
-	}
-#endif
-#ifdef IS_PRALINE
-	if (IS_PRALINE) {
-		enable_rf_power_praline();
-	}
-#endif
-#ifdef IS_RAD1O
-	if (IS_RAD1O) {
-		enable_rf_power_rad1o();
-	}
-#endif
-}
-
-void disable_rf_power(void)
-{
-#ifdef IS_HACKRF_ONE
-	if (IS_HACKRF_ONE) {
-		disable_rf_power_hackrf_one();
-	}
-#endif
-#ifdef IS_PRALINE
-	if (IS_PRALINE) {
-		disable_rf_power_praline();
-	}
-#endif
-#ifdef IS_RAD1O
-	if (IS_RAD1O) {
-		disable_rf_power_rad1o();
-	}
-#endif
 }
 
 void led_on(const led_t led)
