+++ firmware/hackrf_usb/hackrf_usb.c
@@ -38,6 +38,7 @@
 #include <hackrf_ui.h>
 #include <operacake.h>
 #include <platform_detect.h>
+#include <power.h>
 #include <radio.h>
 #include <rf_path.h>
 #include <rom_iap.h>
